Control.Concurrent.Async.Warden

Description

A more flexible way to create Asyncs and have them automatically cancelled when the Warden is shut down.

Documentation

data Warden

A Warden is an owner of Asyncs which cancels them on shutdown.

Nothing in the MVar means the Warden has been shut down.

withWarden :: (Warden -> IO a) -> IO a

Run the action with a new Warden, and call shutdown when the action exits.

create :: IO Warden

Create a new Warden.

shutdown :: Warden -> IO ()

Shutdown a Warden, calling cancel on all owned threads. Subsequent calls to spawn and shutdown will be no-ops.

Note that any exceptions thrown by the threads will be ignored. If you want exceptions to be propagated, either call wait explicitly on the Async, or use link.

spawn :: Warden -> IO a -> IO (Async a)

Spawn a new thread owned by the Warden.

spawn_ :: Warden -> IO () -> IO ()

Spawn a new thread owned by the Warden.

spawnMask :: Warden -> ((forall b. IO b -> IO b) -> IO a) -> IO (Async a)

Spawn a thread with masked exceptions and pass an unmask function to the action.

newtype WardenException

Constructors

WardenException String